feat(ContentFixer): Move post content methods to a new Migrator
